+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 All notable changes to the **AeroSense** project will be documented in this file.
 
+## [v3.15.1] - 2026-03-21
+### Added
+- Modified Interstellar theme.
+- Added sad SpongeBob theme.
+- Added Succession theme.
+
 ## [v3.15.0] - 2026-03-20
 ### Fixed
 - Fixed GUI page refresh on button press clearing terminal messages.